+++ src/lib.rs
@@ -1,6 +1,4 @@
-//! 
-//! No-STD abstraction layer enabling numerical functions to be implemented once, and simultaneously support both fixed and floating point types.
-//! The crate focueses on computationally efficient implementations of numerical operations.
+//! No-STD abstraction layer enabling numerical functions to be implemented once, and simultaneously support both real and complex numbers with, int, fixed and floating point types.
 //! 
 //! This is an experimental library.
 //! 
@@ -21,6 +19,29 @@
 //! 
 //! assert_eq!{ res, 0.5404195 };
 //! ``` 
+//! 
+//! 
+//! The library supplies complex structs `Polar<T>` and `Cartesian<T>` with no-std implementation of math traits, including `MixedNum` traits.
+//! 
+//! Some interoperability with num::Complex is implemented.
+//! 
+//! ## Example
+//! 
+//! ```
+//! use mixed_num::*;
+//! use mixed_num::traits::*;
+//! 
+//! let number = Cartesian::new(1f32,2f32);
+//! 
+//! assert_eq!{ number.to_string(), "1+2i" };
+//! 
+//! let polar_number = number.to_polar();
+//! assert_eq!{ polar_number.to_string(), "2.236068∠1.1071488" };
+//! 
+//! let polar_conj: Polar<f32> = polar_number.conj();
+//! 
+//! assert_eq!{ polar_conj.to_string(), "2.236068∠-1.1071488" };
+//! ``` 
 
 #![crate_name = "mixed_num"]
 #![no_std]
@@ -38,6 +59,10 @@ pub use float_impl::*;
 
 mod fixed_impl;
 pub use fixed_impl::*;
+
+pub mod complex;
+#[allow(unused)]
+pub use complex::*;
 
 mod int_impl;
 pub use int_impl::*;